made __init__.py and moved routes from app.py into routes.py. Also renamed app.py to poke_swap.py. Also added a .flaskenv file and a base.html file
