+++ .flaskenv
@@ -1,0 +1,1 @@
+FLASK_APP=poke_swap.py
